LayoutTests/platform/mac-wk2/accessibility/content-inset-scrollview-frame.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that a page with a content inset will not affect the bounds of the main frame scroll view. It will appear as the same as the web area's sbounds");

    function outputFrame(message, element) {
        debug(message + ": {" + element.x + ", " + element.y + "}, {" + element.width + ", " + element.height + "}");
    }

    if (window.accessibilityController) {
        var webArea = accessibilityController.rootElement.childAtIndex(0);
        debug("Web area role: " + webArea.role);
        outputFrame("WebArea frame", webArea);

        var scrollView = webArea.parentElement();
        debug("ScrollView role: " + scrollView.role);
        outputFrame("ScrollView frame (0 content inset)", scrollView);
    
        window.internals.setTopContentInset(100);

        outputFrame("ScrollView frame (100 content inset)", scrollView);
    }
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
